created utils and services folders for logic purposes
