+++ app/main.py
@@ -3,8 +3,12 @@ from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel
 from typing import List, Optional
 import json
-import pandas as pd # Used to manipulate dataframes
-import Thermobar as TB # Core of calculation
+import pandas as pd
+import Thermobar as TB
+
+from utils.utils import rename_duplicate_columns
+from services.calculations_service import function_constructor
+from utils.models import calculationRequest, calculationResponse
 
 # Create an instance of the app
 appCalculationThermobar = FastAPI()
@@ -17,39 +21,14 @@ appCalculationThermobar.add_middleware(
     allow_headers=["*"],
 )
 
-# Create an object of what the appliction receive
-class CalculationRequest(BaseModel):
-    data: str
-    equationP: Optional[str]
-    equationT: Optional[str]
-    h2o: Optional[float]
-    h2oDependant: bool
-    iterative: bool
-    pDependant: bool
-    phases: List[str]
-    pressure: Optional[float]
-    system: str
-    tDependant: bool
-    temperature: Optional[float]
-
-# Create an object of what the application send back
-class CalculationResponse(BaseModel):
-    phases: List[str]
-    equationP: Optional[str]
-    equationT: Optional[str]
-    pressure: Optional[float]
-    temperature: Optional[float]
-    h2o: Optional[float]
-    data: str
-
 # _________________________________________________________________________________
 
-@appCalculationThermobar.post("/api/calculate", response_model=CalculationResponse)
+@appCalculationThermobar.post("/api/calculate", response_model=calculationResponse)
 
-async def calculate(request: CalculationRequest):
+async def calculate(request: calculationRequest):
 
     phaseConcat = ''
-    phasesVarableList = {}
+    phasesVariableList = {}
     phaseArg = ''
 
     userData = request
@@ -65,7 +44,7 @@ async def calculate(request: CalculationRequest):
 
     for phase in userData.phases:
         globals()[f'compo_{phase.lower()}'] = userData.data.filter(regex = '_' + phase)
-        phasesVarableList[f'compo_{phase.lower()}'] = f'compo_{phase.lower()}'
+        phasesVariableList[f'compo_{phase.lower()}'] = f'compo_{phase.lower()}'
 
         phaseConcat = phaseConcat + '_' + phase.lower() # used later for function name
 
@@ -76,17 +55,17 @@ async def calculate(request: CalculationRequest):
 
     # Creating the function name and the arguments for the different cases ___________________________________
 
-    function_name = functionConstructor(userData.iterative,
-                                        userData.tDependant,
-                                        userData.pDependant,
-                                        userData.h2oDependant,
-                                        userData.equationP,
-                                        userData.equationT,
-                                        phaseArg,
-                                        phaseConcat,
-                                        userData.temperature,
-                                        userData.pressure,
-                                        userData.h2o)
+    function_name = function_constructor(userData.iterative,
+                                         userData.tDependant,
+                                         userData.pDependant,
+                                         userData.h2oDependant,
+                                         userData.equationP,
+                                         userData.equationT,
+                                         phaseArg,
+                                         phaseConcat,
+                                         userData.temperature,
+                                         userData.pressure,
+                                         userData.h2o)
 
     # Get this function using getattr (get attribute of the TB)
     # if the function doesn't exist an error is raised and the system is shut down
@@ -103,7 +82,7 @@ async def calculate(request: CalculationRequest):
 
     calculations = calculations.to_json(orient='records', lines=False)
 
-    return CalculationResponse(
+    return calculationResponse(
 
         phases=request.phases,
         equationP=request.equationP,
@@ -113,74 +92,6 @@ async def calculate(request: CalculationRequest):
         h2o = request.h2o,
         data=calculations
     )
-
-
-
-def functionConstructor(iterative, tDependant, pDependant, h2oDependant, equationP, equationT, phaseArg, phaseConcat, temperature, pressure, h2o):
-
-    if iterative == True:
-
-        # Construct the name of the fonction using the phases provided by the userData object
-        functionName = f'calculate{phaseConcat}_press_temp'
-
-        # Construction of the arguments of the function and function call (arg)
-        eqArg = f'equationP="{equationP}", equationT="{equationT}", '
-        arguments = phaseArg + eqArg + 'eq_tests=True'
-
-    elif equationP != None : # Case to calculate P
-
-        # Construct the name of the fonction using the phases provided by the userData object
-        functionName = f'calculate{phaseConcat}_press, '
-
-        eqArg = f'equationP="{equationP}", '
-        tempArg = ''
-
-        if tDependant == True: # If the equation is temperature dependant the temperature argument is added
-            tempArg = f'T = {temperature}, '
-
-        if h2oDependant == True: # If the equation is water content dependant the water content argument is added
-            h2oArg = f'H2O_Liq = {h2o}, '
-
-        # Construction of the arguments of the function and function call (arg)
-        arguments = phaseArg + eqArg + tempArg + h2oArg + 'eq_tests=True'
-
-    elif equationT != None : # Case to calculate T
-        # Construct the name of the fonction using the phases provided by the userData object
-        functionName = f'calculate{phaseConcat}_temp, '
-
-        eqArg = f'equationT="{equationT}", '
-        pressArg = ''
-
-        if pDependant == True: # If the equation is temperature dependant the temperature argument is added
-            pressArg = f'P = {pressure} ,'
-
-        if h2oDependant == True: # If the equation is water content dependant the water content argument is added
-            h2oArg = f'H2O_Liq = {h2o} ,'
-
-        # Construction of the arguments of the function and function call (arg)
-        arguments = phaseArg + eqArg + pressArg + h2oArg + 'eq_tests=True'
-
-    else :
-        # if no equation is passed (SHOULD NOT HAPPEN) the system shutdown
-        print(f"No equation passed shutting down")
-        raise SystemExit
-
-# Function to correct if there is duplicated column names (which case the change to a json will fail at the end of calculations)
-def rename_duplicate_columns(df):
-
-    new_columns = []
-    column_count = {}
-
-    for col in df.columns:
-        if col in column_count:
-            column_count[col] += 1
-            new_columns.append(f"{col}_{column_count[col]}")
-        else:
-            column_count[col] = 0
-            new_columns.append(col)
-
-    df.columns = new_columns
-    return df
 
 if __name__ == '__main__':
     import uvicorn
